pyproject.toml
[tool.poetry]
name = "mongodb-dump-and-restore"
version = "0.1.0"
description = ""
authors = ["GwangCheonLee <raitmid56@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
motor = "^3.4.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
